Link utests and rename table_builder.rs to ank_table.rs
